% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/just.install.R
\name{install_package}
\alias{install_package}
\title{internal function to that installs a single package based on data provided}
\usage{
install_package(package_name, source_type, source_param)
}
\arguments{
\item{package_name}{package name}

\item{source_type}{where it should be installed from}

\item{source_param}{repository source, where it applies}
}
\value{
no output
}
\description{
internal function to that installs a single package based on data provided
}
