% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/just.install.R
\name{justinstall}
\alias{justinstall}
\title{install packages, only if there are missing and without attaching them}
\usage{
justinstall(to_install)
}
\arguments{
\item{to_install}{tibble or data.frame with packages to install (attr: package: package name, source: {"CRAN","Github","Bioconductor","Other"}, url: username/package for Github, repo url for Other)}
}
\value{
no output
}
\description{
install packages, only if there are missing and without attaching them
}
